pyproject.toml
[tool.poetry]
name = "twitterpersona"
version = "0.3.2"
description = "Assess whether a twitter is positive or negative based on the user's recent tweets"
authors = ["Andy Wang, Renzo Wijngaarden, Roan Raina, Yurui Feng"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tweepy = "^4.12.1"
pandas = "^1.5.2"
python-dotenv = "^0.21.0"
nltk = "^3.8.1"
tweet-preprocessor = "^0.6.0"
matplotlib = "^3.6.3"
pillow = "^9.4.0"
wordcloud = "^1.8.2.2"
numpy = "^1.24.1"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
jupyter = "^1.0.0"
myst-nb = {version = "^0.17.1", python = "^3.9"}
sphinx-autoapi = "^2.0.1"
sphinx-rtd-theme = "^1.1.1"
pytest-cov = "^4.0.0"
python-semantic-release = "^7.33.0"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
